fix(web): keep quiz Canvas sync lazy-loaded to pass bundle budget
Move quiz grade sync helpers out of the shared entry chunk so CI gzip
stays under the entry budget. Restore the inline submission sync waiter in
courses-api to avoid pulling an extra module into the main bundle.
